move opensearch related code into old type
